Added simple classifier

+++ applyMask.m
@@ -1,7 +1,7 @@
 function imgCopy=applyMask(img, mask, coefficient)
     [width, height, cols] = size(img);
     
-    imgCopy = img;
+    imgCopy = double(img);
     for color = 1:cols
         for x = 1:width
             for y = 1:height
